+++ SDL/game/draw.h
@@ -5,7 +5,7 @@
 //In other words, Draw class handles everything that we need to access for drawing stuffs to the surface.
 //Move the SDL_Surface from game, and place it here. Move the texture to the Game.
 
-//TODO(Thompson): In Game class, have the Game class maintain a list of objects to pass to the Draw class, and update the objects individually as needed.
+//TODO(Thompson): In Game class, have the Game class masize_tain a list of objects to pass to the Draw class, and update the objects individually as needed.
 //Draw class should then iterate through the list of updated objects, and render them individually.
 
 #include <SDL.h>
@@ -33,9 +33,9 @@ public:
 	void SetScale(float value);
 
 	//Fetches the pixel value and stores it in the outPixel. Returns true if succeeds. False, if otherwise.
-	bool Draw::GetPixel(int x, int y, uint32_t* outPixel);
-	//Sets the pixel value in the intended position. Returns true if succeeds. False, if otherwise.
-	bool SetPixel(int x, int y, uint32_t color);
+	bool Draw::GetPixel(size_t x, size_t y, uint32_t* outPixel);
+	//Sets the pixel value in the size_tended position. Returns true if succeeds. False, if otherwise.
+	bool SetPixel(size_t x, size_t y, uint32_t color);
 
 	void SetSurface(SDL_Surface* surface);
 	SDL_Surface* GetSurface() const;
